requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==5.1.2
djangorestframework==3.15.2
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
